Add missing build time path environment variables
Ensures that all of the environment variables available at runtime are
also available at build time. This is necessary since it's common for
apps to be invoked at build time (eg for Django collectstatic), and
often this means fully initialising the framework backends, which will
fail if expected libraries are not found.

Fixes #15.
Closes @W-8429195@.