Update default GEOS from 3.7.2 to 3.10.2
This will only affect new apps (or existing apps where the build
cache is cleared) that don't specify an explicit GEOS version, and
that are using Heroku-20 (or the unreleased Heroku-22).

To override the default, set the `GEOS_VERSION` env var to
the desired version.

It was not possible to build GEOS 3.10.2 on Heroku-18, since it
requires CMake 3.13+, which is newer than the CMake available
on Ubuntu 18.04.

Changes:
https://github.com/libgeos/geos/blob/main/NEWS.md#changes-in-3100

GUS-W-10346751.