Add support for PROJ 8.2.1
This is the newest PROJ that doesn't require the updated
CMake based build script (9+ requires CMake, so will require
further changes that aren't worth making for now).

Newer PROJ requires that the sqlite3 headers and binary be
available during the build, so the Dockerfiles had to be updated
accordingly.

The default PROJ remains unchanged for now, so apps wanting
to use newer PROJ will need to set the env var `PROJ_VERSION`
to `8.2.1`.

GUS-W-10346751.