Switch to the `heroku-buildpack-geo` S3 bucket
Previously the assets for this buildpack were stored in an S3 bucket
owned by a Bucketeer addon on a Heroku app, which is inconsistent
with our other buildpacks. (It was set up this way, since the original
buildpack author was on a different team and didn't have access to
the required AWS account).

Now, a new `heroku-buildpack-geo` S3 bucket has been created,
with the existing assets copied over.

GUS-W-11089566.