Include origin of versions in build output
Previously the build output only stated what version of GDAL/GEOS/PROJ
was being installed, and not why that version was chosen.

In particular, this made the "the version you requested doesn't exist
for this stack" message confusing for the case where the user didn't
pick a version, but instead the version was inherited from the previous
build via the cache.

Now, the origin of the versions is logged.

In addition, the "version not available for this stack" message has been
tweaked to not (falsely) suggest the version was always explicitly
requested, plus explain how to adjust the version using the env var or
clearing the build cache.

These changes should make stack upgrades less frustrating (such as for
the upcoming Heroku-24 release).

Lastly, some best practices have been applied to the curl usage.